examples/sqla/requirements.txt
starlette-admin
sqlalchemy